feneffea/requirements.txt
dj-database-url==1.2.0
Django==4.1.7
django-extensions==3.2.3
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.6.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
wagtail==4.2
wagtail-django-recaptcha==1.0